+++ src/components/Form/Form.types.ts
@@ -7,11 +7,15 @@ import type {
   InputType,
   SpecificInputProps,
   InputValue,
+  CommonInputProps,
 } from "@/components/Input";
+import type { ModalProps } from "@/components/Modal";
+import type { GridEndPosition } from "@layouts";
 import type {
   ColorKey,
   StyleOverrideProps,
   StyleProps,
+  TestIdProps,
   TextProps,
 } from "@types";
 import type {
@@ -25,11 +29,13 @@ import type {
   MaybeAsyncFn,
   Nullable,
   Nullish,
+  NullishPrimitives,
   Optional,
   Replace,
   VoidFn,
 } from "@ubloimmo/front-util";
-import type { FormEventHandler, ReactNode } from "react";
+import type { TranslationKey } from "@utils";
+import type { FC, FormEventHandler, ReactNode } from "react";
 import type { ZodIssue, ZodObject, ZodType, ZodTypeAny } from "zod";
 
 // -------------------------------- GLOBALS ----------------------------------
@@ -80,6 +86,31 @@ export type FormData<TData extends object> = DeepPartial<TData>;
 
 export type FormFieldLayoutHiddenFn = GenericFn<[], boolean>;
 
+export type FormFieldLayout = {
+  /**
+   * Horizontal size of the field relative to the form's grid.
+   * Maps to `grid-column-end: span ${size}`
+   *
+   * @remarks if not provided, will default to half of the {@link FormLayoutProps.columns}.
+   *
+   * @default 1
+   */
+  size?: number;
+  /**
+   * Whether to render the field in display mode even while in edit mode
+   *
+   * @default false
+   */
+  readonly?: boolean;
+  /**
+   * Whether to hide the field in both display AND edit modes
+   *
+   * @type {boolean | FormFieldLayoutHiddenFn}
+   * @default false
+   */
+  hidden?: boolean | FormFieldLayoutHiddenFn;
+};
+
 /**
  * Props added to form field that control its layout inside the form
  */
@@ -87,28 +118,17 @@ export type FormFieldLayoutProps = {
   /**
    * Layout of the field
    */
-  layout?: {
-    /**
-     * Horizontal size of the field relative to the form's grid.
-     * Maps to `grid-column-end: span ${size}`
-     *
-     * @default 1
-     */
-    size?: number;
-    /**
-     * Whether to render the field in display mode even while in edit mode
-     *
-     * @default false
-     */
-    readonly?: boolean;
-    /**
-     * Whether to hide the field in both display AND edit modes
-     *
-     * @type {boolean | FormFieldLayoutHiddenFn}
-     * @default false
-     */
-    hidden?: boolean | FormFieldLayoutHiddenFn;
-  };
+  layout?: FormFieldLayout;
+};
+
+export type BuiltFormFieldLayout = Required<
+  Replace<FormFieldLayout, "hidden", { hidden: boolean }>
+> & {
+  columnEnd: GridEndPosition;
+};
+
+export type BuiltFormFieldLayoutProps = {
+  layout: BuiltFormFieldLayout;
 };
 
 /**
@@ -207,7 +227,8 @@ export type FormFieldProps<TData extends object> = {
  *
  */
 export type BuiltFieldProps<TType extends InputType> = FieldProps<TType> &
-  FormFieldLayoutProps;
+  BuiltFormFieldLayoutProps &
+  TestIdProps;
 
 // ------------------------------- DIVIDER ----------------------------------
 
@@ -217,6 +238,9 @@ export type FormDividerProps =
 
 // -------------------------------- INFOS -----------------------------------
 
+/**
+ * Props to pass to a `Text` component
+ */
 export type FormTextProps = Omit<
   TextProps,
   "children" | keyof StyleOverrideProps
@@ -229,6 +253,80 @@ export type BuiltFormTextProps = Omit<TextProps, keyof StyleOverrideProps> & {
   kind: "text";
 };
 
+// --------------------------- CUSTOM CONTENT -------------------------------
+
+/**
+ * Abritraty content, either an object marked as `kind: "content"` or a React functional component
+ */
+export type FormCustomContentProps =
+  | {
+      content: ReactNode | FC;
+      kind: "content";
+    }
+  | FC;
+
+export type BuiltFormCustomContentProps = Exclude<FormCustomContentProps, FC>;
+
+// ----------------------------- CUSTOM FIELDS ------------------------------
+
+export type CustomFormInputProps<TValue extends NullishPrimitives> =
+  CommonInputProps & {
+    /**
+     * The custom field's value or null if empty
+     *
+     * @default null
+     *
+     * @type {InputValue | null}
+     *
+     */
+    value?: Nullable<TValue>;
+    /**
+     * The custom field's onChange callback. Optional.
+     *
+     * @default null
+     *
+     * @type {InputOnChangeFn | null}
+     */
+    onChange?: Nullable<VoidFn<[Nullable<TValue>]>>;
+
+    /**
+     * The custom field's name
+     *
+     * @default null
+     */
+    name?: Nullable<string>;
+  };
+
+type PreservedFieldProps = Omit<
+  FieldProps<InputType>,
+  Exclude<
+    keyof CustomFormInputProps<NullishPrimitives>,
+    "onChange" | "disabled" | "error" | "required"
+  >
+>;
+
+export type FormCustomFieldProps<TData extends object> = {
+  [TSource in FormSource<TData>]: DeepValueOf<
+    CompleteFormData<TData>,
+    TSource
+  > extends infer TFieldValue
+    ? FormFieldLayoutProps &
+        PreservedFieldProps & {
+          kind: "custom-field";
+          source: TSource;
+          CustomInput: FC<
+            CustomFormInputProps<TFieldValue & NullishPrimitives>
+          >;
+        }
+    : never;
+}[FormSource<TData>];
+
+export type BuiltFormCustomFieldProps = PreservedFieldProps &
+  BuiltFormFieldLayoutProps &
+  CustomFormInputProps<NullishPrimitives> & {
+    CustomInput: FC<CustomFormInputProps<NullishPrimitives>>;
+  };
+
 // ------------------------------- CONTENT ----------------------------------
 
 /**
@@ -240,8 +338,10 @@ export type BuiltFormTextProps = Omit<TextProps, keyof StyleOverrideProps> & {
  */
 export type FormContent<TData extends object> =
   | FormFieldProps<TData>
+  | FormCustomFieldProps<TData>
   | FormDividerProps
-  | FormTextProps;
+  | FormTextProps
+  | FormCustomContentProps;
 
 /**
  * Either a single {@link BuiltFieldProps} object or {@link FormDividerProps}
@@ -250,7 +350,9 @@ export type FormContent<TData extends object> =
  */
 export type BuiltFormContent<TType extends InputType> =
   | BuiltFieldProps<TType>
+  | BuiltFormCustomFieldProps
   | FormDividerProps
+  | FormCustomContentProps
   | BuiltFormTextProps;
 
 // -------------------------------- SCHEMA ----------------------------------
@@ -304,6 +406,15 @@ export type FormOnSubmitFn<TData extends object> = MaybeAsyncFn<
   boolean | FormData<TData> | void
 >;
 
+/**
+ * Form `onSubmitError` callback
+ *
+ * Gets triggered when an errors occurs during form submission
+ *
+ * @param {Error} error - The error that occurred
+ */
+export type FormOnSubmitErrorFn = VoidFn<[Error]>;
+
 // --------------------------------- PROPS ----------------------------------
 
 /**
@@ -332,6 +443,36 @@ export type FormHeaderProps = {
    */
   icon?: Nullable<IconName>;
 };
+
+export type FormModalProps = {
+  /**
+   * Optional {@link ModalProps} to display in a modal
+   *
+   * @remarks providing valid Modal props will render the form in a `Modal` component.
+   *
+   * @type {Nullable<Omit<ModalProps, "title">>}
+   * @default null
+   */
+  asModal?: Nullable<Omit<ModalProps, "title">>;
+};
+
+export type FormGridProps = {
+  /**
+   * Optional number of columns to display the form's content in
+   *
+   * @remarks Odd values will be rounded up to even values
+   *
+   * @type {number}
+   * @default 2
+   */
+  columns?: number;
+};
+
+export type FormLayoutProps = FormGridProps & FormModalProps;
+
+export type DefaultFormGridProps = Required<FormGridProps>;
+
+export type DefaultFormLayoutProps = Required<FormLayoutProps>;
 
 /**
  * Form props related to its content
@@ -389,6 +530,17 @@ export type FormDataProps<TData extends object> = {
    * Callback called when form is submitted
    */
   onSubmit?: Nullable<FormOnSubmitFn<TData>>;
+  /**
+   * Callback called when an errors occurs during form submission
+   */
+  onSubmitError?: Nullable<FormOnSubmitErrorFn>;
+  /**
+   * Callback called when form edition is cancelled
+   *
+   * @type {Nullable<VoidFn>}
+   * @default null
+   */
+  onCancelled?: Nullable<VoidFn>;
 };
 
 /**
@@ -448,6 +600,28 @@ export type FormModifierProps = {
 };
 
 /**
+ * Form props related to the banner displayed in `edit` mode
+ */
+export type FormEditBannerProps = {
+  /**
+   * Label of the submit button
+   *
+   * @type {TranslationKey<"action">}
+   * @default "save"
+   */
+  submitLabel?: TranslationKey<"action">;
+  /**
+   * Label of the cancel button
+   *
+   * @type {TranslationKey<"action">}
+   * @default "cancel"
+   */
+  cancelLabel?: TranslationKey<"action">;
+};
+
+export type DefaultFormEditBannerProps = Required<FormEditBannerProps>;
+
+/**
  * Form props directly passed to its underlying context
  *
  * @template {object} TData - The type of the form data
@@ -466,6 +640,8 @@ export type FormContextProps<TData extends object> = FormDataProps<TData> &
  * @see {@link FormHeaderProps}, {@link FormContextProps}
  */
 export type FormProps<TData extends object> = FormHeaderProps &
+  FormLayoutProps &
+  FormEditBannerProps &
   FormContextProps<TData>;
 
 /**
@@ -544,6 +720,21 @@ export type MutateFormDataFn<TData extends object> = <
 export type BuildFieldPropsFn<TData extends object> = <TType extends InputType>(
   formField: FormFieldProps<TData>
 ) => BuiltFieldProps<TType>;
+
+/**
+ * Builds a {@link BuiltFormCustomFieldProps} object from a {@link FormCustomFieldProps} object
+ * links its `value`, `onChange`, `disabled`, `required`, `error` and `errorText` properties with the form data & schema,
+ * while preserving its `CustomInput` component;
+ *
+ * @template {object} TData - The type of the {@link FormData}
+ *
+ * @param {FormCustomFieldProps<TData>} formCustomField - The {@link FormCustomFieldProps}
+ *
+ * @returns {BuiltFormCustomFieldProps<TData>} The built and linked {@link BuiltFormCustomFieldProps}
+ */
+export type BuildCustomFieldPropsFn<TData extends object> = (
+  formCustomField: FormCustomFieldProps<TData>
+) => BuiltFormCustomFieldProps;
 
 /**
  * Retrieves the value of a {@link FormField} from the internal {@link FormData} using its `source`.
@@ -628,6 +819,11 @@ export type IsFieldRequiredFn<TData extends object> = (
  * @returns {FormValidation} The form's validation state
  */
 export type ComputeFormValidationFn = GenericFn<[], FormValidation>;
+
+export type BuildFormFieldLayoutFn = GenericFn<
+  [Optional<FormFieldLayout>],
+  BuiltFormFieldLayoutProps["layout"]
+>;
 
 // -------------------------- DISPLAY TRANSFORMS -----------------------------
 
@@ -796,6 +992,11 @@ export type UseFormEditStateReturn = {
   stopEditing: VoidFn;
 };
 
+export type UseFormLayoutReturn = {
+  columns: number;
+  buildFormFieldLayout: BuildFormFieldLayoutFn;
+};
+
 // ------------------------------- CONTEXT ---------------------------------
 
 /**
@@ -811,7 +1012,8 @@ export type FormContext<TData extends object> = UseFormDataReturn<TData> &
   UseFormValidationReturn<TData> &
   UseFormSubmissionReturn &
   FormModifers &
-  UseFormEditStateReturn & {
+  UseFormEditStateReturn &
+  UseFormLayoutReturn & {
     /**
      * An array of {@link BuiltFieldProps},
      * used to render the form's fields