+++ src/components/Input/IconPickerInput/IconPickerInput.component.tsx
@@ -1,10 +1,11 @@
 import { useCallback } from "react";
 
 import { defaultCommonInputProps } from "../Input.common";
+import { useInputId } from "../Input.utils";
 
 import { IconPicker } from "@/components/IconPicker";
 import { TestIdProps } from "@types";
-import { useTestId } from "@utils";
+import { useMergedProps, useTestId } from "@utils";
 
 import type {
   IconPickerInputDefaultProps,
@@ -36,19 +37,23 @@ const defaultIconPickerInputProps: IconPickerInputDefaultProps = {
 const IconPickerInput = (
   props: IconPickerInputProps & TestIdProps
 ): JSX.Element => {
+  const mergedProps = useMergedProps(defaultIconPickerInputProps, props);
   const testId = useTestId("input-icon-picker", props);
 
   const onIconPickerChange = useCallback<IconPickerInputOnChangeFn>(
     (value) => {
-      if (!props.onChange) return;
-      return props.onChange(value);
+      if (!mergedProps.onChange) return;
+      return mergedProps.onChange(value);
     },
-    [props]
+    [mergedProps]
   );
+
+  const id = useInputId(mergedProps);
 
   return (
     <IconPicker
-      {...props}
+      {...mergedProps}
+      id={id}
       testId={testId}
       overrideTestId
       onChange={onIconPickerChange}