+++ src/layouts/Flex/Flex.types.ts
@@ -96,6 +96,13 @@ export type FlexLayoutProps = {
    * @type {AriaRole}
    */
   role?: Nullable<AriaRole>;
+  /**
+   * An optional id for the element
+   *
+   * @type {Nullable<string>}
+   * @default null
+   */
+  id?: Nullable<string>;
 } & StyleOverrideProps;
 
 export type FlexLayoutDefaultProps = Required<FlexLayoutProps>;